feat(model): add theme type to theme entity
