+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management/gravitee-apim-rest-api-management-rest/src/main/java/io/gravitee/rest/api/management/rest/resource/ThemeResource.java
@@ -21,8 +21,8 @@ import io.gravitee.rest.api.model.PictureEntity;
 import io.gravitee.rest.api.model.UrlPictureEntity;
 import io.gravitee.rest.api.model.permissions.RolePermission;
 import io.gravitee.rest.api.model.permissions.RolePermissionAction;
-import io.gravitee.rest.api.model.theme.ThemeEntity;
-import io.gravitee.rest.api.model.theme.UpdateThemeEntity;
+import io.gravitee.rest.api.model.theme.portal.ThemeEntity;
+import io.gravitee.rest.api.model.theme.portal.UpdateThemeEntity;
 import io.gravitee.rest.api.rest.annotation.Permission;
 import io.gravitee.rest.api.rest.annotation.Permissions;
 import io.gravitee.rest.api.service.ThemeService;
