+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/ThemeServiceImpl.java
@@ -31,7 +31,12 @@ import io.gravitee.repository.management.model.Theme;
 import io.gravitee.rest.api.model.InlinePictureEntity;
 import io.gravitee.rest.api.model.PictureEntity;
 import io.gravitee.rest.api.model.UrlPictureEntity;
-import io.gravitee.rest.api.model.theme.*;
+import io.gravitee.rest.api.model.theme.portal.NewThemeEntity;
+import io.gravitee.rest.api.model.theme.portal.ThemeComponentDefinition;
+import io.gravitee.rest.api.model.theme.portal.ThemeCssDefinition;
+import io.gravitee.rest.api.model.theme.portal.ThemeDefinition;
+import io.gravitee.rest.api.model.theme.portal.ThemeEntity;
+import io.gravitee.rest.api.model.theme.portal.UpdateThemeEntity;
 import io.gravitee.rest.api.service.AuditService;
 import io.gravitee.rest.api.service.ThemeService;
 import io.gravitee.rest.api.service.common.ExecutionContext;
